functions/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azurefunctions-extensions-bindings-blob
tensorflow==2.18.0
onnxmltools
#ai-edge-litert==1.0.1
numpy
pillow